feat: 뷰모델 값 이동 및 코드 정리
