+++ Projects/Feature/ReportFeature/Sources/Presentation/Model/ReportModel.swift
@@ -21,8 +21,10 @@ extension ReportModel {
          ReportModel(title: "제보할 때 어떤 정보를 제공해야 하나요?", subtitle: """
                                                                                                     다음 정보를 제공해 주시면 감사합니다:
                                                                                                     
-                                                                                                    1. 핵 의심 장면의 스크린샷
-                                                                                                    2. 핵 의심 유저의 닉네임 또는 병영 번호
+                                                                                                    - 닉네임 및 병영번호
+                                                                                                    - 스크린샷
+                                                                                                    - 날짜 및 시간
+                                                                                                    - 상세 설명
                                                                                                     """),
          ReportModel(title: "제보한 내용은 어떻게 처리되나요?", subtitle: "제보된 내용은 저희 팀이 검토하여 핵 사용 여부를 판단합니다."),
          ReportModel(title: "제보 결과를 확인할 수 있나요?", subtitle: "제보 결과는 개인정보 보호를 위해 공개되지 않습니다. 제보된 유저가 핵 사용자로 판명될 경우 적절한 조치가 취해집니다."),
